orchestrator/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
nats-py==2.6.0
python-dotenv==1.0.1

